重构 Agent 运行时模块
